feat(issues): implement issue management endpoints; add getAllIssues and addIssue functions
