+++ src/routes/visits.route.js
@@ -1,5 +1,5 @@
 import { Router } from "express";
-import { addIssue, createVisit, getAllIssues, getAllVisitsCount, getCancelledVisits, getPendingVisitsCount, getSuccessfulVisits, getUpcomingVisits, getVisitById, updateVisit, updateVisitNotes, updateVisitStaff, updateVisitStatus } from "../controller/visits.controller.js";
+import { createVisit, getAllVisitsCount, getCancelledVisits, getPendingVisitsCount, getSuccessfulVisits, getUpcomingVisits, getVisitById, updateVisit, updateVisitNotes, updateVisitStaff, updateVisitStatus } from "../controller/visits.controller.js";
 import { verifyJWT } from "../middleware/auth.middleware.js";
 
 const router = Router();
@@ -11,11 +11,9 @@ router.get('/get-pending-visits-count', verifyJWT, getPendingVisitsCount);
 router.get('/get-successful-visits', verifyJWT, getSuccessfulVisits);
 router.get('/get-cancelled-visits', verifyJWT, getCancelledVisits)
 router.get('/get-visit/:id', verifyJWT, getVisitById)
-router.get('/get-all-issues/:id', verifyJWT, getAllIssues)
 router.patch('/update-visit/:id', verifyJWT, updateVisit);
 router.patch('/update-visit-status/:id', verifyJWT, updateVisitStatus);
 router.patch('/update-visit-staff/:id', verifyJWT, updateVisitStaff);
 router.patch('/update-visit-notes/:id', verifyJWT, updateVisitNotes);
-router.patch('/update-add-issue/:id', verifyJWT, addIssue)
 
 export default router;
